+++ react-front-end/src/components/BookingCalendar.js
@@ -1,0 +1,42 @@
+import React from "react";
+import { Calendar, momentLocalizer } from 'react-big-calendar';
+import moment from 'moment';
+import './MyCalendarStyle.scss';
+import MyCalendar from "./MyCalendar"
+
+const localizer = momentLocalizer(moment);
+
+const BookingCalendar = () => {
+  return (
+    <div>
+
+      {/* Launch demo modal */}
+      <button type="button" className="btn btn-primary" data-bs-toggle="modal" data-bs-target="#exampleModal">
+      </button>
+
+      <div className="modal fade" id="exampleModal" tabIndex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
+        <div className="modal-dialog">
+          <div className="modal-content">
+            <div className="modal-header">
+              <h1 className="modal-title fs-5" id="exampleModalLabel">Modal title</h1>
+              <button type="button" className="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+            </div>
+            <div className="modal-body">
+
+
+              <MyCalendar localizer={localizer} />
+
+
+            </div>
+            <div className="modal-footer">
+              <button type="button" className="btn btn-secondary" data-bs-dismiss="modal">Close</button>
+              {/* <button type="button" className="btn btn-primary">Save changes</button> */}
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default BookingCalendar;
