--- react-front-end/src/components/CalendarStyle.scss
@@ -1,10 +1,0 @@
-// CalendarStyles.scss
-@import 'react-big-calendar/lib/sass/styles';
-@import 'react-big-calendar/lib/addons/dragAndDrop/styles'; // if using DnD
-
-$bc-primary-color: #007bff; // Change the primary color to blue
-
-// Additional custom styles
-.myCustomHeight {
-  height: 800px; // Set a custom height for the calendar container
-}
