+++ react-front-end/src/components/Home.js
@@ -7,7 +7,8 @@ import TutorCard from "./TutorCard";
 import { useEffect } from "react";
 import TutorPage from "./TutorPage"
 import './Home.scss';
-import Calendar from "./Calendar"
+import MyCalendar from "./MyCalendar"
+import BookingCalendar from "./BookingCalendar"
 // import 'bootstrap/dist/css/bootstrap.min.css';
 
 
@@ -66,28 +67,7 @@ const Home = () => {
         ))}
       </ul>
 
-      <button type="button" class="btn btn-primary" data-bs-toggle="modal" data-bs-target="#exampleModal">
-  {/* Launch demo modal */}
-      </button>
-
-    {/* Modal */}
-    <div class="modal fade" id="exampleModal" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
-      <div class="modal-dialog">
-        <div class="modal-content">
-          <div class="modal-header">
-            <h1 class="modal-title fs-5" id="exampleModalLabel">Modal title</h1>
-            <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-          </div>
-          <div class="modal-body">
-            <Calendar />
-          </div>
-          <div class="modal-footer">
-            <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
-            <button type="button" class="btn btn-primary">Save changes</button>
-          </div>
-        </div>
-      </div>
-    </div>
+      <BookingCalendar />
 
       <div className="tutor-container">
         {tutors.map((tutor) => (
