+++ react-front-end/src/index.js
@@ -19,7 +19,8 @@ import StudentPage from "./components/StudentPage";
 import StudentMessages from "./components/StudentMessages";
 import StudentBookings from "./components/StudentBookings";
 import LessonPlanWriter from "./components/LessonPlanWriter";
-import Calendar from "./components/Calendar";
+import BookingCalendar from "./components/BookingCalendar";
+import MyCalendar from "./components/MyCalendar";
 import 'bootstrap/dist/css/bootstrap.min.css';
 import 'bootstrap/dist/js/bootstrap.min.js';
 
@@ -44,7 +45,8 @@ export default function Index() {
         <Route path="/StudentMessages" element={<StudentMessages />} /> 
         <Route path="/StudentBookings" element={<StudentBookings />} /> 
         <Route path="/LessonPlanWriter" element={<LessonPlanWriter />} />
-        <Route path="/Calendar" element={<Calendar />} /> 
+        <Route path="/BookingCalendar" element={<BookingCalendar />} />
+        <Route path="/MyCalendar" element={<MyCalendar />} />  
 
       </Routes>
     </BrowserRouter>
